# openapi.yml
openapi: 3.1.0
info:
  title: API de réservation de salles
  version: 1.0.0
  description: |
    Contrat de l'API de gestion des salles, équipements, organisateurs et réservations.
    Les dates et heures utilisent le format ISO 8601 avec fuseau ou décalage UTC.
servers:
  - url: http://localhost:8080
    description: Environnement local
...
paths:
...
  /api/reservations:
    post:
      tags: [Reservations]
      summary: Réserver une salle précise
      description: |
        Le client choisit la salle avec `roomId`. La réservation est confirmée uniquement
        si la salle et l'organisateur existent, si la salle est `AVAILABLE`, si sa capacité
        est suffisante, si elle possède tous les équipements demandés et si aucune
        réservation confirmée ne chevauche la période demandée.

        Le début doit être dans le futur et strictement antérieur à la fin. La durée
        maximale est de huit heures. Une réservation consécutive à une autre est admise.
        Si la salle demandée est incompatible, elle n'est jamais remplacée automatiquement.
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Réservation confirmée
          headers:
            Location:
              description: URI de la réservation créée
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/ReservationBadRequest'
        '404':
          $ref: '#/components/responses/ReservationResourceNotFound'
        '409':
          $ref: '#/components/responses/ReservationConflict'
    get:
      tags: [Reservations]
      summary: Lister et filtrer les réservations
      description: |
        Retourne les réservations confirmées et annulées. Tous les filtres fournis sont
        cumulés. Avec `from` et `to`, sont retenues les réservations qui chevauchent la
        période demandée. Le résultat est trié par date de début puis par identifiant.
      operationId: listReservations
      parameters:
        - name: roomId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ResourceId'
        - name: organizerId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ResourceId'
        - name: from
          in: query
          required: false
          description: Inclut les réservations dont la fin est strictement postérieure à cette date.
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: Inclut les réservations dont le début est strictement antérieur à cette date.
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Réservations correspondant aux filtres, triées par début puis identifiant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/ReservationQueryBadRequest'
  /api/reservations/automatic:
    post:
      tags: [Reservations]
      summary: Réserver automatiquement la salle la plus adaptée
      description: |
        Le client ne fournit pas de salle. Après validation de l'organisateur et de la
        période, le système ne conserve que les salles `AVAILABLE` ayant une capacité
        suffisante, tous les équipements demandés et aucune réservation confirmée en
        conflit avec la période.

        Pour chaque candidate :

        - `unusedCapacity = room.capacity - numberOfParticipants` ;
        - dans le même bâtiment (`room.building.id = organizer.building.id`),
          `distance = abs(room.floor - organizer.floor)` ;
        - dans des bâtiments différents,
          `distance = 10 + abs(room.floor - organizer.floor)` ;
        - `score = distance * 10 + unusedCapacity`.

        Le score le plus faible est le meilleur : un étage de distance équivaut à dix places
        inutilisées, tandis qu'un changement de bâtiment ajoute une pénalité de 100
        points. Les égalités sont départagées par nom de salle sans tenir compte de la
        casse, puis par identifiant. La première salle est attribuée et la réservation
        est créée de manière atomique. Si aucune salle n'est compatible, aucune
        réservation n'est créée et l'erreur `NO_COMPATIBLE_ROOM` est renvoyée.
      operationId: createAutomaticReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutomaticReservationRequest'
      responses:
        '201':
          description: Réservation confirmée avec la salle attribuée
          headers:
            Location:
              description: URI de la réservation créée
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/AutomaticReservationBadRequest'
        '404':
          $ref: '#/components/responses/AutomaticReservationResourceNotFound'
        '409':
          $ref: '#/components/responses/NoCompatibleRoom'
  /api/reservations/{reservationId}:
    parameters:
      - $ref: '#/components/parameters/ReservationId'
    get:
      tags: [Reservations]
      summary: Consulter une réservation
      description: Retourne la réservation demandée, qu'elle soit confirmée ou annulée.
      operationId: getReservation
      responses:
        '200':
          description: Réservation trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '404':
          $ref: '#/components/responses/ReservationNotFound'
  /api/reservations/{reservationId}/cancel:
    parameters:
      - $ref: '#/components/parameters/ReservationId'
    patch:
      tags: [Reservations]
      summary: Annuler une réservation
      description: |
        Passe une réservation `CONFIRMED` au statut `CANCELLED`. La réservation reste
        consultable et ne bloque plus la salle. Une seconde annulation renvoie
        `RESERVATION_ALREADY_CANCELLED`.
      operationId: cancelReservation
      responses:
        '200':
          description: Réservation annulée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '404':
          $ref: '#/components/responses/CancellationReservationNotFound'
        '409':
          $ref: '#/components/responses/ReservationAlreadyCancelled'
components:
  parameters:
...
    ReservationId:
      name: reservationId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ResourceId'
...
  responses:
...
    ReservationBadRequest:
      description: Les données ou la période de réservation sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            invalidPeriod:
              summary: Période invalide
              value:
                code: INVALID_RESERVATION_PERIOD
                message: Une réservation ne peut pas dépasser huit heures
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details: {}
                fieldErrors: {}
            validationError:
              summary: Champ invalide
              value:
                code: VALIDATION_ERROR
                message: La requête contient des données invalides
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details: {}
                fieldErrors:
                  numberOfParticipants: doit être strictement positif
    AutomaticReservationBadRequest:
      description: Les données ou la période de réservation automatique sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            invalidPeriod:
              summary: Période invalide
              value:
                code: INVALID_RESERVATION_PERIOD
                message: Une réservation ne peut pas dépasser huit heures
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations/automatic
                details: {}
                fieldErrors: {}
            validationError:
              summary: Champ invalide
              value:
                code: VALIDATION_ERROR
                message: La requête contient des données invalides
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations/automatic
                details: {}
                fieldErrors:
                  numberOfParticipants: doit être strictement positif
    ReservationQueryBadRequest:
      description: Les filtres de recherche sont invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: INVALID_RESERVATION_PERIOD
            message: Le paramètre from doit être antérieur au paramètre to
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/reservations
            details: {}
            fieldErrors: {}
...
    ReservationNotFound:
      description: La réservation demandée n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: RESERVATION_NOT_FOUND
            message: La réservation 42 n'existe pas
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/reservations/42
            details:
              reservationId: 42
            fieldErrors: {}
    CancellationReservationNotFound:
      description: La réservation à annuler n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: RESERVATION_NOT_FOUND
            message: La réservation 42 n'existe pas
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/reservations/42/cancel
            details:
              reservationId: 42
            fieldErrors: {}
...
    ReservationResourceNotFound:
      description: La salle, l'organisateur ou un équipement demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            roomNotFound:
              value:
                code: ROOM_NOT_FOUND
                message: La salle 7 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  roomId: 7
                fieldErrors: {}
            organizerNotFound:
              value:
                code: ORGANIZER_NOT_FOUND
                message: L'organisateur 12 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  organizerId: 12
                fieldErrors: {}
            equipmentNotFound:
              value:
                code: EQUIPMENT_NOT_FOUND
                message: L'équipement PROJECTOR n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  equipmentCode: PROJECTOR
                fieldErrors: {}
    AutomaticReservationResourceNotFound:
      description: L'organisateur ou un équipement demandé n'existe pas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            organizerNotFound:
              value:
                code: ORGANIZER_NOT_FOUND
                message: L'organisateur 12 n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations/automatic
                details:
                  organizerId: 12
                fieldErrors: {}
            equipmentNotFound:
              value:
                code: EQUIPMENT_NOT_FOUND
                message: L'équipement PROJECTOR n'existe pas
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations/automatic
                details:
                  equipmentCode: PROJECTOR
                fieldErrors: {}
...
    ReservationConflict:
      description: La salle choisie ne peut pas accueillir la réservation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          examples:
            roomUnavailable:
              value:
                code: ROOM_UNAVAILABLE
                message: La salle Orion est en maintenance
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  roomId: 7
                fieldErrors: {}
            capacityExceeded:
              value:
                code: ROOM_CAPACITY_EXCEEDED
                message: La capacité de la salle est insuffisante
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  roomId: 7
                  roomCapacity: 20
                  numberOfParticipants: 25
                fieldErrors: {}
            missingEquipment:
              value:
                code: MISSING_REQUIRED_EQUIPMENT
                message: La salle ne possède pas tous les équipements demandés
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  roomId: 7
                  missingEquipmentCodes: [PROJECTOR]
                fieldErrors: {}
            roomAlreadyReserved:
              value:
                code: ROOM_ALREADY_RESERVED
                message: La salle Orion est déjà réservée sur cette période
                timestamp: '2026-10-01T13:45:12Z'
                path: /api/reservations
                details:
                  roomId: 7
                  conflictingReservationId: 38
                fieldErrors: {}
    NoCompatibleRoom:
      description: Aucune salle n'est compatible avec les critères de réservation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: NO_COMPATIBLE_ROOM
            message: Aucune salle disponible ne correspond aux critères demandés
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/reservations/automatic
            details: {}
            fieldErrors: {}
    ReservationAlreadyCancelled:
      description: La réservation est déjà annulée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: RESERVATION_ALREADY_CANCELLED
            message: La réservation 42 est déjà annulée
            timestamp: '2026-10-01T13:45:12Z'
            path: /api/reservations/42/cancel
            details:
              reservationId: 42
            fieldErrors: {}
  schemas:
    ResourceId:
      type: integer
      format: int64
      minimum: 1
      example: 7
    RoomStatus:
      type: string
      enum: [AVAILABLE, MAINTENANCE]
    ReservationStatus:
      type: string
      enum: [CONFIRMED, CANCELLED]
    EquipmentCode:
      type: string
      pattern: '^[A-Z][A-Z0-9_]{1,49}$'
      example: PROJECTOR
    BuildingName:
      type: string
      minLength: 1
      maxLength: 100
      description: Nom unique du bâtiment, sans tenir compte de la casse
      example: Bâtiment A
    FloorNumber:
      type: integer
      format: int32
      minimum: 0
      maximum: 199
      description: Doit être inférieur au nombre d'étages du bâtiment référencé
      example: 2
    NumberOfFloors:
      type: integer
      format: int32
      minimum: 1
      maximum: 200
      description: Nombre d'étages, rez-de-chaussée inclus
      example: 5
...
    BuildingResponse:
      type: object
      additionalProperties: false
      required: [id, name, numberOfFloors]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          $ref: '#/components/schemas/BuildingName'
        numberOfFloors:
          $ref: '#/components/schemas/NumberOfFloors'
...
    RoomSummaryResponse:
      type: object
      required: [id, name, building, floor, capacity, status]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          example: Orion
        building:
          $ref: '#/components/schemas/BuildingResponse'
        floor:
          $ref: '#/components/schemas/FloorNumber'
        capacity:
          type: integer
          format: int32
          minimum: 1
          example: 30
        status:
          $ref: '#/components/schemas/RoomStatus'
...
    OrganizerSummaryResponse:
      type: object
      required: [id, name, building, floor]
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        name:
          type: string
          example: Alice Martin
        building:
          $ref: '#/components/schemas/BuildingResponse'
        floor:
          $ref: '#/components/schemas/FloorNumber'
...
    ReservationRequestFields:
      type: object
      required: [title, organizerId, start, end, numberOfParticipants]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Soutenance de projet
        organizerId:
          $ref: '#/components/schemas/ResourceId'
        start:
          type: string
          format: date-time
          example: '2026-10-15T14:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-10-15T16:00:00+02:00'
        numberOfParticipants:
          type: integer
          format: int32
          minimum: 1
          example: 25
        requiredEquipmentCodes:
          type: array
          uniqueItems: true
          default: []
          items:
            $ref: '#/components/schemas/EquipmentCode'
          example: [PROJECTOR, VIDEO_CONFERENCE]
    CreateReservationRequest:
      allOf:
        - $ref: '#/components/schemas/ReservationRequestFields'
        - type: object
          required: [roomId]
          properties:
            roomId:
              $ref: '#/components/schemas/ResourceId'
    AutomaticReservationRequest:
      allOf:
        - $ref: '#/components/schemas/ReservationRequestFields'
    ReservationResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - status
        - room
        - organizer
        - start
        - end
        - numberOfParticipants
        - requiredEquipmentCodes
        - createdAt
      properties:
        id:
          $ref: '#/components/schemas/ResourceId'
        title:
          type: string
          example: Soutenance de projet
        status:
          $ref: '#/components/schemas/ReservationStatus'
        room:
          $ref: '#/components/schemas/RoomSummaryResponse'
        organizer:
          $ref: '#/components/schemas/OrganizerSummaryResponse'
        start:
          type: string
          format: date-time
          example: '2026-10-15T14:00:00+02:00'
        end:
          type: string
          format: date-time
          example: '2026-10-15T16:00:00+02:00'
        numberOfParticipants:
          type: integer
          format: int32
          minimum: 1
          example: 25
        requiredEquipmentCodes:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/EquipmentCode'
        createdAt:
          type: string
          format: date-time
          example: '2026-10-01T13:45:12Z'
...
    ApiErrorResponse:
      type: object
      description: |
        Format commun à toutes les erreurs de l'API.

        | Code | HTTP | Signification |
        | --- | ---: | --- |
        | `VALIDATION_ERROR` | 400 | Un ou plusieurs champs sont invalides. |
        | `INVALID_RESERVATION_PERIOD` | 400 | La période est passée, inversée ou dépasse huit heures. |
        | `BUILDING_NOT_FOUND` | 404 | Le bâtiment demandé n'existe pas. |
        | `ROOM_NOT_FOUND` | 404 | La salle demandée n'existe pas. |
        | `ORGANIZER_NOT_FOUND` | 404 | L'organisateur demandé n'existe pas. |
        | `RESERVATION_NOT_FOUND` | 404 | La réservation demandée n'existe pas. |
        | `EQUIPMENT_NOT_FOUND` | 404 | Au moins un code d'équipement n'existe pas. |
        | `ROOM_CAPACITY_EXCEEDED` | 409 | La salle choisie est trop petite. |
        | `MISSING_REQUIRED_EQUIPMENT` | 409 | La salle choisie ne possède pas tous les équipements demandés. |
        | `ROOM_ALREADY_RESERVED` | 409 | Une réservation confirmée chevauche la période. |
        | `ROOM_UNAVAILABLE` | 409 | La salle choisie est en maintenance. |
        | `NO_COMPATIBLE_ROOM` | 409 | Aucune salle ne convient à l'attribution automatique. |
        | `RESERVATION_ALREADY_CANCELLED` | 409 | La réservation est déjà annulée. |
        | `RESOURCE_ALREADY_EXISTS` | 409 | Un nom, code ou e-mail unique est déjà utilisé. |
        | `BUILDING_FLOOR_COUNT_CONFLICT` | 409 | Le nombre d'étages est incompatible avec une localisation existante. |
      additionalProperties: false
      required: [code, message, timestamp, path]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - INVALID_RESERVATION_PERIOD
            - BUILDING_NOT_FOUND
            - ROOM_NOT_FOUND
            - ORGANIZER_NOT_FOUND
            - RESERVATION_NOT_FOUND
            - EQUIPMENT_NOT_FOUND
            - ROOM_CAPACITY_EXCEEDED
            - MISSING_REQUIRED_EQUIPMENT
            - ROOM_ALREADY_RESERVED
            - ROOM_UNAVAILABLE
            - NO_COMPATIBLE_ROOM
            - RESERVATION_ALREADY_CANCELLED
            - RESOURCE_ALREADY_EXISTS
            - BUILDING_FLOOR_COUNT_CONFLICT
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object
          additionalProperties: true
          default: {}
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          default: {}